RUB-248 rust: split tx helper signing shape
